requirements.txt
##################
# News Challenge #
##################

## System dependencies and utilities
gunicorn>=19.3,<19.4

## Data Analytics
pandas>=0.18,<0.19
django-pandas>=0.4,<0.5

## Django
Django>=1.9,<1.10
django-extensions>=1.6,<1.7
psycopg2>=2.6,<2.7
django-material>=0.7,<1.0

## Django REST Framework
djangorestframework>=3.3,<3.4

## Cache and Redis
django-redis>=4.3,<4.4

## IPython
ipython>=4.0,<4.1

## Tests
ipdb
coverage==3.7.1
